+++ app/controllers/intents_controller.rb
@@ -5,8 +5,15 @@ class IntentsController < ApplicationController
   prepend_before_action :set_draw
 
   def report
+    @intents_importer = IntentsImportForm.new(students: nil)
     @students_by_intent = UsersByIntentQuery.new(@draw.students).call
     @intent_metrics = @students_by_intent.transform_values(&:count)
+  end
+
+  def import
+    result = IntentsImportForm.import(students: @draw.students,
+                                      file: intents_import_params['file'].path)
+    handle_action(path: report_draw_intents_path(@draw), **result)
   end
 
   def export
@@ -24,6 +31,10 @@ class IntentsController < ApplicationController
   end
 
   def set_draw
-    @draw = Draw.find(params[:draw_id])
+    @draw = Draw.includes(:students).find(params[:draw_id])
+  end
+
+  def intents_import_params
+    params.require(:intents_import_form).permit(:file)
   end
 end